+++ common/src/main/java/dev/tocraft/walkers/mixin/player/PlayerEntityDataMixin.java
@@ -12,7 +12,6 @@ import dev.tocraft.walkers.traits.impl.RiderTrait;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.resources.Identifier;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
@@ -20,7 +19,6 @@ import net.minecraft.world.entity.EntitySpawnReason;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeInstance;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.animal.equine.AbstractHorse;
 import net.minecraft.world.entity.player.Player;
@@ -211,7 +209,7 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
         AttributeInstance armorToughnessAttribute = player.getAttribute(Attributes.ARMOR_TOUGHNESS);
 
         // spawn particles
-        if (Walkers.CONFIG.emit_particles && !Objects.equals(ShapeType.from(this.walkers$shape), ShapeType.from(shape)) && player.level() instanceof ServerLevel l) {
+        if (Walkers.CONFIG.emit_particles && !Objects.equals(ShapeType.from(this.walkers$shape), ShapeType.from(shape)) && !player.isSpectator() && player.level() instanceof ServerLevel l) {
             l.sendParticles(ParticleTypes.TOTEM_OF_UNDYING, player.getX(), player.getY() + 1, player.getZ(), 25, .5, 1.0, .5, .75);
             l.sendParticles(ParticleTypes.PORTAL, player.getX(), player.getY() + 1, player.getZ(), 25, .5, 1.0, .5, .75);
             l.sendParticles(ParticleTypes.ELECTRIC_SPARK, player.getX(), player.getY() + 1, player.getZ(), 25, .5, 1.0, .5, .7);